Refactor: RightSidebar 부분 리팩토링
