+++ src/pages/main/components/RightSidebar/ButtonOptions/ButtonnOptionsContainer.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const ButtonOptionsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+
+  gap: 24px;
+`;
